refactor: rename engine to prompt
