+++ src/prompt/image_test.go
@@ -1,4 +1,4 @@
-package engine
+package prompt
 
 import (
 	"os"
