+++ src/prompt/prompt.go
@@ -1,4 +1,4 @@
-package engine
+package prompt
 
 import (
 	"fmt"
